% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dist_stop_times.R
\name{calculate_dist_stop_times}
\alias{calculate_dist_stop_times}
\title{Calculate distances and speeds for stop_times segments}
\usage{
calculate_dist_stop_times(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A \code{gtfs} object with updated stop_times
}
\description{
Returns the GTFS with the stop_time file with columns of distance and speed
}
\section{Details}{

Details here.
}

\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")

gtfs <- read_gtfs(data_path)

trip_geometry <- get_trip_geometry(gtfs)
head(trip_geometry)

# the above is identical to
trip_geometry <- get_trip_geometry(gtfs, file = c("shapes", "stop_times"))
head(trip_geometry)

trip_ids <- c("CPTM L07-0", "2002-10-0")
trip_geometry <- get_trip_geometry(gtfs, trip_id = trip_ids)
trip_geometry
plot(trip_geometry["origin_file"])

}
