% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_route_frequency.R
\name{calculate_route_frequency}
\alias{calculate_route_frequency}
\title{Calculate route frequency}
\usage{
calculate_route_frequency(
  gtfs,
  route_id = NULL,
  service_id = NULL,
  start_time = "06:00:00",
  end_time = "08:00:00"
)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{route_id}{A character specifying a route (from \code{routes}) to subset, if desired}

\item{start_time}{A character specifying the start of the window}

\item{end_time}{A character specifying the end of the window}

\item{service_id1}{A character specifying a service_id (from \code{calendar}) to subset, if desired}

\item{mean_headway}{TRUE/FASE - Should return the mean headway in the entire window (TRUE) or the hourly mean headway (FALSE)?}
}
\value{
A \code{data.table} with the headway for each \code{route} and \code{shape_id}
}
\description{
Calculate headway between departures for route and direction
}
