% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_to_sf.R
\name{shapes_to_sf}
\alias{shapes_to_sf}
\title{Ecreate linestrings from shapes}
\usage{
shapes_to_sf(gtfs, shape_id = NULL)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{service_id}{A character specifying a service (from the \code{calendar}) to subset, if desired. This is specially
useful for large GTFS file where one is interested only on weekdays trips, for example.}

\item{route_id}{A character specifying a route (from the \code{routes}) to subset, if desired}
}
\value{
A \code{data.table}.with the scheduled service (sequence of stops) for each \code{route} and \code{shape_id}
}
\description{
Returns the most common service (the sequence of stops on the \code{stop_times} file) for each shape
}
\section{Details}{

Details here.
}

\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")

gtfs <- read_gtfs(data_path)

trip_geometry <- get_trip_geometry(gtfs)
head(trip_geometry)

# the above is identical to
trip_geometry <- get_trip_geometry(gtfs, file = c("shapes", "stop_times"))
head(trip_geometry)

trip_ids <- c("CPTM L07-0", "2002-10-0")
trip_geometry <- get_trip_geometry(gtfs, trip_id = trip_ids)
trip_geometry
plot(trip_geometry["origin_file"])

}
