% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataio.R
\name{write_data}
\alias{write_data}
\title{Write data based on my structure}
\usage{
write_data(data, file)
}
\arguments{
\item{data}{Data to be written to disk}

\item{file}{Filename}
}
\description{
Write data based on my structure
}
