% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataio.R
\name{read_data}
\alias{read_data}
\title{Read data based on my structure}
\usage{
read_data(file, type = "data")
}
\arguments{
\item{file}{Filename}

\item{type}{Type. data or data-raw?}
}
\description{
Read data based on my structure
}
